+++ RazorForms/Templates/Elements/ErrorList.cshtml
@@ -1,4 +1,4 @@
-﻿@inherits RazorForms.Templates.TemplateBase<RazorForms.FormInput>
+﻿@inherits RazorForms.Templates.Elements.ErrorListBase<RazorForms.FormInput>
 
 @if (Model.Errors.Any())
 {
