+++ RazorForms/Templates/Elements/Label.cshtml
@@ -1,4 +1,4 @@
-﻿@inherits RazorForms.Templates.TemplateBase<RazorForms.FormInput>
+﻿@inherits RazorForms.Templates.Elements.LabelBase<RazorForms.FormInput>
 
 @if (Model.Options.RemoveWrappers ?? false)
 {
