+++ RazorForms/Templates/Elements/LabelRaw.cshtml
@@ -1,4 +1,4 @@
-﻿@inherits RazorForms.Templates.TemplateBase<RazorForms.FormInput>
+﻿@inherits RazorForms.Templates.Elements.LabelRawBase<RazorForms.FormInput>
 
 <label for="@Model.MvcName"
        class="@GenerateLabelClasses()">
