+++ RazorForms/Templates/Inputs/Input.cshtml
@@ -1,4 +1,4 @@
-﻿@inherits RazorForms.Templates.TemplateBase<RazorForms.FormInput>
+﻿@inherits RazorForms.Templates.Inputs.InputBase<RazorForms.FormInput>
 
 @if (Model.Options.RemoveWrappers ?? false)
 {
