+++ RazorForms/Templates/Inputs/InputRaw.cshtml
@@ -1,6 +1,6 @@
 ﻿@using RazorForms
 @using InputType = RazorForms.InputType
-@inherits RazorForms.Templates.TemplateBase<RazorForms.FormInput>
+@inherits RazorForms.Templates.Inputs.InputRawBase<RazorForms.FormInput>
 
 @if (Model.Type == InputType.TextArea)
 {
