+++ RazorForms/Templates/TemplateBase.cs
@@ -7,54 +7,6 @@ namespace RazorForms.Templates;
 public abstract class TemplateBase<TModel> : RazorPage<TModel>
 	where TModel : FormInput
 {
-	protected string GenerateComponentWrapperClasses()
-		=> GenerateClasses(Model.Options.ComponentWrapperClasses,
-		                   Model.Options.ComponentWrapperValidClasses,
-		                   Model.Options.ComponentWrapperErrorClasses);
-
-	protected string GenerateInputBlockWrapperClasses()
-		=> GenerateClasses(Model.Options.InputBlockWrapperClasses,
-		                   Model.Options.InputBlockWrapperValidClasses,
-		                   Model.Options.InputBlockWrapperErrorClasses);
-
-	protected string GenerateLabelWrapperClasses()
-		=> GenerateClasses(Model.Options.LabelWrapperClasses,
-		                   Model.Options.LabelWrapperValidClasses,
-		                   Model.Options.LabelWrapperErrorClasses);
-
-	protected string GenerateLabelClasses()
-		=> GenerateClasses(Model.Options.LabelClasses,
-		                   Model.Options.LabelValidClasses,
-		                   Model.Options.LabelErrorClasses);
-
-	protected string GenerateInputWrapperClasses()
-		=> GenerateClasses(Model.Options.InputWrapperClasses,
-		                   Model.Options.InputWrapperValidClasses,
-		                   Model.Options.InputWrapperErrorClasses);
-
-	protected string GenerateInputClasses()
-	{
-		var classAttribute = Model.Attributes.FirstOrDefault(a => a.Name == "class");
-		var classValue = classAttribute?.Value.ToString() ?? string.Empty;
-
-		return GenerateClasses(Model.Options.InputClasses,
-		                       Model.Options.InputValidClasses,
-		                       Model.Options.InputErrorClasses,
-		                       startingValue: classValue);
-	}
-
-	protected string GenerateErrorWrapperClasses()
-		=> GenerateClasses(Model.Options.ErrorWrapperClasses,
-		                   string.Empty,
-		                   string.Empty,
-		                   false);
-
-	protected string GenerateErrorClasses()
-		=> GenerateClasses(Model.Options.ErrorClasses,
-		                   string.Empty,
-		                   string.Empty,
-		                   false);
-
 	protected string GenerateAttributes()
 	{
 		var sb = new StringBuilder();
@@ -94,7 +46,11 @@ public abstract class TemplateBase<TModel> : RazorPage<TModel>
 
 	protected static string GenerateErrorId(string input) => $"{input}Errors";
 
-	private string GenerateClasses(string? defaultClass = null, string? validStateClass = null, string? invalidStateClass = null, bool checkValidationState = true, string? startingValue = null)
+	protected string GenerateClasses(string? defaultClass = null,
+	                                 string? validStateClass = null,
+	                                 string? invalidStateClass = null,
+	                                 bool checkValidationState = true,
+	                                 string? startingValue = null)
 	{
 		var classes = new StringBuilder(startingValue);
 		classes.Append(' ');
