+++ RazorForms/Templates/TextInput.cshtml
@@ -1,5 +1,5 @@
 ﻿@using RazorForms
-@inherits RazorForms.Templates.TemplateBase<FormInput>
+@inherits RazorForms.Templates.TextInputBase<FormInput>
 
 <div class="@GenerateComponentWrapperClasses()">
 	<div class="@GenerateInputBlockWrapperClasses()">
